Move reprepro functions to a lib module.
