+++ lib/archive_auth_mirror/scripts/mirror_archive.py
@@ -1,31 +1,22 @@
-import subprocess
+"""Mirror and update a repository."""
 
 from ..utils import get_paths
+from ..reprepro import reprepro
 from ..script import (
     setup_logger,
     get_config,
 )
 
 
-def reprepro(*args):
-    """Call reprepro with the specified args."""
-    paths = get_paths()
-    command = [
-        'reprepro',
-        '--basedir', str(paths['reprepro']),
-        '--confdir', str(paths['reprepro-conf']),
-        '--outdir', str(paths['static'] / 'ubuntu'),
-        '--gnupghome', str(paths['gnupghome'])]
-    command.extend(args)
-    subprocess.check_call(command)
-
-
 def main():
     logger = setup_logger()
     paths = get_paths()
-    config = get_config(paths['config'])
+    suite = get_config(paths['config'])['suite']
 
     logger.info('starting mirroring')
-    reprepro('--show-percent', 'update', config['suite'])
+    reprepro(
+        '--show-percent', '--export=never', '--keepunreferencedfiles',
+        'update', suite)
+    reprepro('export', suite)
     reprepro('deleteunreferenced')
     logger.info('mirroring completed')
